+++ task_9/index.html
@@ -53,6 +53,6 @@
     </div>
   </footer>
 
-  <script src="js/script.js"></script>
+  <script src="js/script.js" type="module"></script>
 </body>
 </html>
